+++ lib/IRGen/GenHeap.cpp
@@ -30,6 +30,7 @@
 
 #include "ConstantBuilder.h"
 #include "Explosion.h"
+#include "GenClass.h"
 #include "GenProto.h"
 #include "GenType.h"
 #include "IRGenDebugInfo.h"
@@ -37,6 +38,7 @@
 #include "IRGenModule.h"
 #include "HeapTypeInfo.h"
 #include "IndirectTypeInfo.h"
+#include "MetadataRequest.h"
 #include "WeakTypeInfo.h"
 
 #include "GenHeap.h"
@@ -1785,3 +1787,196 @@ DEFINE_COPY_OP(UnknownWeakCopyInit)
 DEFINE_COPY_OP(UnknownWeakCopyAssign)
 DEFINE_COPY_OP(UnknownWeakTakeInit)
 DEFINE_COPY_OP(UnknownWeakTakeAssign)
+
+llvm::Value *IRGenFunction::getLocalSelfMetadata() {
+  assert(LocalSelf && "no local self metadata");
+  switch (SelfKind) {
+  case SwiftMetatype:
+    return LocalSelf;
+  case ObjCMetatype:
+    return emitObjCMetadataRefForMetadata(*this, LocalSelf);
+  case ObjectReference:
+    return emitDynamicTypeOfOpaqueHeapObject(*this, LocalSelf);
+  }
+
+  llvm_unreachable("Not a valid LocalSelfKind.");
+}
+
+/// Given a non-tagged object pointer, load a pointer to its class object.
+llvm::Value *irgen::emitLoadOfObjCHeapMetadataRef(IRGenFunction &IGF,
+                                                  llvm::Value *object) {
+  if (IGF.IGM.TargetInfo.hasISAMasking()) {
+    object = IGF.Builder.CreateBitCast(object,
+                                       IGF.IGM.IntPtrTy->getPointerTo());
+    llvm::Value *metadata =
+      IGF.Builder.CreateLoad(Address(object, IGF.IGM.getPointerAlignment()));
+    llvm::Value *mask = IGF.Builder.CreateLoad(IGF.IGM.getAddrOfObjCISAMask());
+    metadata = IGF.Builder.CreateAnd(metadata, mask);
+    metadata = IGF.Builder.CreateIntToPtr(metadata, IGF.IGM.TypeMetadataPtrTy);
+    return metadata;
+  } else if (IGF.IGM.TargetInfo.hasOpaqueISAs()) {
+    return emitHeapMetadataRefForUnknownHeapObject(IGF, object);
+  } else {
+    object = IGF.Builder.CreateBitCast(object,
+                                  IGF.IGM.TypeMetadataPtrTy->getPointerTo());
+    llvm::Value *metadata =
+      IGF.Builder.CreateLoad(Address(object, IGF.IGM.getPointerAlignment()));
+    return metadata;
+  }
+}
+
+/// Given a pointer to a heap object (i.e. definitely not a tagged
+/// pointer), load its heap metadata pointer.
+static llvm::Value *emitLoadOfHeapMetadataRef(IRGenFunction &IGF,
+                                              llvm::Value *object,
+                                              IsaEncoding isaEncoding,
+                                              bool suppressCast) {
+  switch (isaEncoding) {
+  case IsaEncoding::Pointer: {
+    // Drill into the object pointer.  Rather than bitcasting, we make
+    // an effort to do something that should explode if we get something
+    // mistyped.
+    llvm::StructType *structTy =
+      cast<llvm::StructType>(
+        cast<llvm::PointerType>(object->getType())->getElementType());
+
+    llvm::Value *slot;
+
+    // We need a bitcast if we're dealing with an opaque class.
+    if (structTy->isOpaque()) {
+      auto metadataPtrPtrTy = IGF.IGM.TypeMetadataPtrTy->getPointerTo();
+      slot = IGF.Builder.CreateBitCast(object, metadataPtrPtrTy);
+
+    // Otherwise, make a GEP.
+    } else {
+      auto zero = llvm::ConstantInt::get(IGF.IGM.Int32Ty, 0);
+
+      SmallVector<llvm::Value*, 4> indexes;
+      indexes.push_back(zero);
+      do {
+        indexes.push_back(zero);
+
+        // Keep drilling down to the first element type.
+        auto eltTy = structTy->getElementType(0);
+        assert(isa<llvm::StructType>(eltTy) || eltTy == IGF.IGM.TypeMetadataPtrTy);
+        structTy = dyn_cast<llvm::StructType>(eltTy);
+      } while (structTy != nullptr);
+
+      slot = IGF.Builder.CreateInBoundsGEP(object, indexes);
+
+      if (!suppressCast) {
+        slot = IGF.Builder.CreateBitCast(slot,
+                                    IGF.IGM.TypeMetadataPtrTy->getPointerTo());
+      }
+    }
+
+    auto metadata = IGF.Builder.CreateLoad(Address(slot,
+                                               IGF.IGM.getPointerAlignment()));
+    if (IGF.IGM.EnableValueNames && object->hasName())
+      metadata->setName(llvm::Twine(object->getName()) + ".metadata");
+    return metadata;
+  }
+      
+  case IsaEncoding::ObjC: {
+    // Feed the object pointer to object_getClass.
+    llvm::Value *objcClass = emitLoadOfObjCHeapMetadataRef(IGF, object);
+    objcClass = IGF.Builder.CreateBitCast(objcClass, IGF.IGM.TypeMetadataPtrTy);
+    return objcClass;
+  }
+  }
+
+  llvm_unreachable("Not a valid IsaEncoding.");
+}
+
+/// Given an object of class type, produce the heap metadata reference
+/// as an %objc_class*.
+llvm::Value *irgen::emitHeapMetadataRefForHeapObject(IRGenFunction &IGF,
+                                                     llvm::Value *object,
+                                                     CanType objectType,
+                                                     bool suppressCast) {
+  ClassDecl *theClass = objectType.getClassOrBoundGenericClass();
+  if (theClass && isKnownNotTaggedPointer(IGF.IGM, theClass))
+    return emitLoadOfHeapMetadataRef(IGF, object,
+                                     getIsaEncodingForType(IGF.IGM, objectType),
+                                     suppressCast);
+
+  // OK, ask the runtime for the class pointer of this potentially-ObjC object.
+  return emitHeapMetadataRefForUnknownHeapObject(IGF, object);
+}
+
+llvm::Value *irgen::emitHeapMetadataRefForHeapObject(IRGenFunction &IGF,
+                                                     llvm::Value *object,
+                                                     SILType objectType,
+                                                     bool suppressCast) {
+  return emitHeapMetadataRefForHeapObject(IGF, object,
+                                          objectType.getSwiftRValueType(),
+                                          suppressCast);
+}
+
+/// Given an opaque class instance pointer, produce the type metadata reference
+/// as a %type*.
+llvm::Value *irgen::emitDynamicTypeOfOpaqueHeapObject(IRGenFunction &IGF,
+                                                      llvm::Value *object) {
+  object = IGF.Builder.CreateBitCast(object, IGF.IGM.ObjCPtrTy);
+  auto metadata = IGF.Builder.CreateCall(IGF.IGM.getGetObjectTypeFn(),
+                                         object,
+                                         object->getName() + ".Type");
+  metadata->setDoesNotThrow();
+  metadata->setOnlyReadsMemory();
+  return metadata;
+}
+
+llvm::Value *irgen::
+emitHeapMetadataRefForUnknownHeapObject(IRGenFunction &IGF,
+                                        llvm::Value *object) {
+  object = IGF.Builder.CreateBitCast(object, IGF.IGM.ObjCPtrTy);
+  auto metadata = IGF.Builder.CreateCall(IGF.IGM.getGetObjectClassFn(),
+                                         object,
+                                         object->getName() + ".Type");
+  metadata->setCallingConv(llvm::CallingConv::C);
+  metadata->setDoesNotThrow();
+  metadata->addAttribute(llvm::AttributeList::FunctionIndex,
+                         llvm::Attribute::ReadOnly);
+  return metadata;
+}
+
+/// Given an object of class type, produce the type metadata reference
+/// as a %type*.
+llvm::Value *irgen::emitDynamicTypeOfHeapObject(IRGenFunction &IGF,
+                                                llvm::Value *object,
+                                                SILType objectType,
+                                                bool suppressCast) {
+  // If it is known to have swift metadata, just load.
+  if (hasKnownSwiftMetadata(IGF.IGM, objectType.getSwiftRValueType())) {
+    return emitLoadOfHeapMetadataRef(IGF, object,
+                getIsaEncodingForType(IGF.IGM, objectType.getSwiftRValueType()),
+                suppressCast);
+  }
+
+  // Okay, ask the runtime for the type metadata of this
+  // potentially-ObjC object.
+  return emitDynamicTypeOfOpaqueHeapObject(IGF, object);
+}
+
+static ClassDecl *getRootClass(ClassDecl *theClass) {
+  while (theClass->hasSuperclass()) {
+    theClass = theClass->getSuperclass()->getClassOrBoundGenericClass();
+    assert(theClass && "base type of class not a class?");
+  }
+  return theClass;
+}
+
+/// What isa encoding mechanism does a type have?
+IsaEncoding irgen::getIsaEncodingForType(IRGenModule &IGM,
+                                         CanType type) {
+  if (auto theClass = type->getClassOrBoundGenericClass()) {
+    // We can access the isas of pure Swift classes directly.
+    if (getRootClass(theClass)->hasKnownSwiftImplementation())
+      return IsaEncoding::Pointer;
+    // For ObjC or mixed classes, we need to use object_getClass.
+    return IsaEncoding::ObjC;
+  }
+  // Non-class heap objects should be pure Swift, so we can access their isas
+  // directly.
+  return IsaEncoding::Pointer;
+}
