+++ lib/IRGen/GenProto.h
@@ -128,7 +128,11 @@ namespace irgen {
                                  Explosion &args,
                                  WitnessMetadata *witnessMetadata,
                                  const GetParameterFn &getParameter);
-  
+ 
+  void emitPolymorphicParametersFromArray(IRGenFunction &IGF,
+                                          NominalTypeDecl *typeDecl,
+                                          Address array);
+
   /// When calling a polymorphic call, pass the arguments for the
   /// generics clause.
   void emitPolymorphicArguments(IRGenFunction &IGF,
@@ -148,6 +152,16 @@ namespace irgen {
   void emitWitnessTableRefs(IRGenFunction &IGF, const Substitution &sub,
                             llvm::Value **metadataCache,
                             SmallVectorImpl<llvm::Value *> &out);
+
+  /// \brief Load a reference to the protocol descriptor for the given protocol.
+  ///
+  /// For Swift protocols, this is a constant reference to the protocol
+  /// descriptor symbol.
+  /// For ObjC protocols, descriptors are uniqued at runtime by the ObjC
+  /// runtime. We need to load the unique reference from a global variable fixed up at
+  /// startup.
+  llvm::Value *emitProtocolDescriptorRef(IRGenFunction &IGF,
+                                         ProtocolDecl *protocol);
 
   /// Emit a witness table reference.
   llvm::Value *emitWitnessTableRef(IRGenFunction &IGF,
