+++ lib/IRGen/GenThunk.cpp
@@ -19,7 +19,8 @@
 #include "Callee.h"
 #include "Explosion.h"
 #include "GenDecl.h"
-#include "GenMeta.h"
+#include "GenClass.h"
+#include "GenHeap.h"
 #include "GenOpaque.h"
 #include "GenProto.h"
 #include "IRGenFunction.h"
