+++ lib/IRGen/MetadataRequest.h
@@ -18,14 +18,24 @@
 #define SWIFT_IRGEN_METADATAREQUEST_H
 
 #include "swift/ABI/MetadataValues.h"
+#include "swift/AST/Types.h"
 
 namespace llvm {
-  class Value;
+class Constant;
+class Function;
+class GlobalVariable;
+class Value;
 }
 
 namespace swift {
+enum ForDefinition_t : bool;
+
 namespace irgen {
-  class IRGenFunction;
+class ConstantReference;
+class Explosion;
+class IRGenFunction;
+class IRGenModule;
+enum class SymbolReferenceKind : unsigned char;
 
 /// A dynamic metadata request.
 class DynamicMetadataRequest {
@@ -107,6 +117,131 @@ public:
   /// (MetadataRequest::Complete).
   static llvm::Constant *getCompletedState(IRGenModule &IGM);
 };
+
+enum class MetadataAccessStrategy {
+  /// There is a unique public accessor function for the given type metadata.
+  PublicUniqueAccessor,
+
+  /// There is a unique hidden accessor function for the given type metadata.
+  HiddenUniqueAccessor,
+
+  /// There is a unique private accessor function for the given type metadata.
+  PrivateAccessor,
+
+  /// There is no unique accessor function for the given type metadata, but
+  /// one should be made automatically.
+  NonUniqueAccessor
+};
+
+/// Is it basically trivial to access the given metadata?  If so, we don't
+/// need a cache variable in its accessor.
+bool isTypeMetadataAccessTrivial(IRGenModule &IGM, CanType type);
+
+/// Determine how the given type metadata should be accessed.
+MetadataAccessStrategy getTypeMetadataAccessStrategy(CanType type);
+
+/// Return the address of a function that will return type metadata 
+/// for the given non-dependent type.
+llvm::Function *getOrCreateTypeMetadataAccessFunction(IRGenModule &IGM,
+                                                      CanType type);
+
+llvm::Function *getTypeMetadataAccessFunction(IRGenModule &IGM,
+                                              CanType type,
+                                              ForDefinition_t shouldDefine);
+
+using MetadataAccessGenerator =
+  llvm::function_ref<MetadataResponse(IRGenFunction &IGF,
+                                      DynamicMetadataRequest request,
+                                      llvm::Constant *cache)>;
+
+llvm::Function *getTypeMetadataAccessFunction(IRGenModule &IGM,
+                                              CanType type,
+                                              ForDefinition_t shouldDefine,
+                                             MetadataAccessGenerator generator);
+
+llvm::Function *
+getGenericTypeMetadataAccessFunction(IRGenModule &IGM,
+                                     NominalTypeDecl *nominal,
+                                     ForDefinition_t shouldDefine);
+
+llvm::Constant *
+getRequiredTypeMetadataAccessFunction(IRGenModule &IGM,
+                                      NominalTypeDecl *theDecl,
+                                      ForDefinition_t shouldDefine);
+
+using InPlaceMetadataInitializer =
+  llvm::function_ref<llvm::Value*(IRGenFunction &IGF, llvm::Value *metadata)>;
+
+MetadataResponse
+emitInPlaceTypeMetadataAccessFunctionBody(IRGenFunction &IGF,
+                                          CanNominalType type,
+                                          llvm::Constant *cacheVariable,
+                                    InPlaceMetadataInitializer &&initializer);
+
+llvm::Value *uniqueForeignTypeMetadataRef(IRGenFunction &IGF,
+                                          llvm::Value *candidate);
+
+llvm::Constant *getTypeRef(IRGenModule &IGM, CanType type);
+
+using LazyCacheEmitter =
+  llvm::function_ref<MetadataResponse(IRGenFunction &IGF, Explosion &params)>;
+
+/// Emit the body of a lazy cache access function.
+void emitLazyCacheAccessFunction(IRGenModule &IGM,
+                                 llvm::Function *accessor,
+                                 llvm::GlobalVariable *cache,
+                                 LazyCacheEmitter getValue,
+                                 bool isReadNone = true);
+
+/// Emit a declaration reference to a metatype object.
+void emitMetatypeRef(IRGenFunction &IGF, CanMetatypeType type,
+                     Explosion &explosion);
+
+/// Emit a reference to a compile-time constant piece of type metadata, or
+/// return a null pointer if the type's metadata cannot be represented by a
+/// constant.
+ConstantReference tryEmitConstantTypeMetadataRef(IRGenModule &IGM,
+                                                 CanType type,
+                                                 SymbolReferenceKind refKind);
+
+/// Get the type as it exists in Swift's runtime type system, removing any
+/// erased generic parameters.
+CanType getRuntimeReifiedType(IRGenModule &IGM, CanType type);
+
+/// Emit a reference to a compile-time constant piece of heap metadata, or
+/// return a null pointer if the type's heap metadata cannot be represented
+/// by a constant.
+llvm::Constant *tryEmitConstantHeapMetadataRef(IRGenModule &IGM,
+                                               CanType type,
+                                               bool allowUninitialized);
+
+enum class MetadataValueType { ObjCClass, TypeMetadata };
+
+/// Emit a reference to the heap metadata for a class.
+///
+/// \returns a value of type ObjCClassPtrTy or TypeMetadataPtrTy,
+///    depending on desiredType
+llvm::Value *emitClassHeapMetadataRef(IRGenFunction &IGF, CanType type,
+                                      MetadataValueType desiredType,
+                                      bool allowUninitialized = false);
+
+/// Emit a reference to the (initialized) ObjC heap metadata for a class.
+///
+/// \returns a value of type ObjCClassPtrTy
+llvm::Value *emitObjCHeapMetadataRef(IRGenFunction &IGF, ClassDecl *theClass,
+                                     bool allowUninitialized = false);
+
+/// Emit a reference to type metadata corresponding to the given
+/// heap metadata.  This may be ObjCWrapper metadata if the heap metadata
+/// is not a class.
+llvm::Value *emitObjCMetadataRefForMetadata(IRGenFunction &IGF,
+                                            llvm::Value *classPtr);
+
+/// Given a class metadata reference, produce the appropriate heap
+/// metadata reference for it.
+llvm::Value *emitClassHeapMetadataRefForMetatype(IRGenFunction &IGF,
+                                                 llvm::Value *metatype,
+                                                 CanType type);
 
 } // end namespace irgen
 } // end namespace swift
